requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langsmith==0.1.144
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
orjson==3.10.11
openai==1.3.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
psutil==6.1.0
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.7.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==12.0
yarl==1.18.0
zipp==3.21.0
